move all code to sandscastle demo
