+++ packages/sandcastle/gallery/3d-tiles-footprint-generator-dev/index.html
@@ -49,8 +49,8 @@
   <label style="color: #fff; font-size: 13px; padding: 4px 8px">
     Clip Quality:
     <select id="clipQualitySelect">
-      <option value="0.25" selected>Low</option>
-      <option value="0.5">Medium</option>
+      <option value="0.25">Low</option>
+      <option value="0.5" selected>Medium</option>
       <option value="1.0">High</option>
     </select>
   </label>
